Clean up database handling
