+++ src/recipe/database.rs
@@ -1,0 +1,7 @@
+pub mod recipe_overview;
+pub mod recipe_details;
+pub mod recipe_ingredients_view;
+
+pub use recipe_overview::RecipeOverview;
+pub use recipe_details::RecipeDetails;
+pub use recipe_ingredients_view::RecipeIngredientsView;
